api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
weaviate-client==4.6.5
markdown-it-py==3.0.0
sentence-transformers==3.0.1
rank-bm25==0.2.2
numpy==1.26.4
regex==2024.7.24
python-dotenv==1.0.1